pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "trafilatura>=2.0.0",
    "mcp>=0.1.0",
    "fastmcp>=2.13.1",
    "groq>=0.36.0",
    "streamlit>=1.51.0",
]
